Refactor monaco initializing
