+++ frontend/src/components/MonacoEditor.tsx
@@ -8,23 +8,7 @@ import { Button, useTheme } from '@material-ui/core'
 import SaveDialog from './SaveDialog'
 import AuthenticateDialog from './AuthenticateDialog'
 import { RepoStateQueryResult } from '../types'
-
-// import { loadVSCodeOnigurumWASM } from '../utils/monacoUtils'
-
-import type { LanguageId } from '../utils/languages'
-import type {
-  ScopeName,
-  TextMateGrammar,
-  ScopeNameInfo,
-} from '../utils/providers'
-import { createOnigScanner, createOnigString } from 'vscode-oniguruma'
-import { SimpleLanguageInfoProvider } from '../utils/providers'
-import { registerLanguages } from '../utils/languages'
-import { rehydrateRegexps } from '../utils/configuration'
-import VsCodeDarkTheme from '../themes/vs-dark-plus-theme'
-import { IOnigLib } from 'vscode-textmate'
-import { grammar } from '../grammars/robot'
-import { robotConfiguration } from '../grammars/robotConfiguration'
+import { initMonaco } from '../utils/monacoInitializer'
 
 interface Props {
   content: string | undefined
@@ -38,10 +22,6 @@ interface Getter {
 interface DialogError {
   title: string
   message: string
-}
-
-interface DemoScopeNameInfo extends ScopeNameInfo {
-  path: string
 }
 
 const MonacoEditor = ({ content, filename }: Props) => {
@@ -112,60 +92,8 @@ const MonacoEditor = ({ content, filename }: Props) => {
     setDialogOpen(true)
   }
 
-  interface Language {
-    id: string
-    extensions: string[]
-  }
-
   monaco.init().then(async (monaco) => {
-    const languages = [
-      {
-        id: 'robot',
-        extensions: ['.robot'],
-      },
-    ]
-
-    const grammars: { [scopeName: string]: DemoScopeNameInfo } = {
-      'source.robot': {
-        language: 'robot',
-        path: 'robotframework.tmLanguage.json',
-      },
-    }
-
-    const fetchGrammar = async (): Promise<any> => {
-      const type = 'json'
-      const json = JSON.stringify(grammar)
-      return { type, grammar: json }
-    }
-
-    const fetchConfiguration = async (): Promise<any> => {
-      const rawConfiguration = JSON.stringify(robotConfiguration)
-      return rehydrateRegexps(rawConfiguration)
-    }
-
-    const onigLib = Promise.resolve({
-      createOnigScanner,
-      createOnigString,
-    })
-
-    const provider = new SimpleLanguageInfoProvider({
-      grammars,
-      fetchGrammar,
-      configurations: languages.map((language: any) => language.id),
-      fetchConfiguration,
-      theme: VsCodeDarkTheme,
-      onigLib,
-      monaco,
-    })
-    registerLanguages(
-      languages,
-      (language: LanguageId) => provider.fetchLanguageInfo(language),
-      monaco
-    )
-
-    monaco.editor.setTheme('vs-dark')
-
-    provider.injectCSS()
+    initMonaco(monaco)
   })
 
   return (
@@ -210,30 +138,6 @@ const MonacoEditor = ({ content, filename }: Props) => {
       </div>
     </div>
   )
-}
-
-function getSampleCodeForLanguage(language: LanguageId): string {
-  if (language === 'robot') {
-    return `\
-*** Settings ***
-Documentation     A test suite with a single test for valid login.
-...
-...               This test has a workflow that is created using keywords in
-...               the imported resource file.
-Resource          resource.txt
-
-*** Test Cases ***
-Valid Login
-    Open Browser To Login Page
-    Input Username    demo
-    Input Password    mode
-    Submit Credentials
-    Welcome Page Should Be Open
-    [Teardown]    Close Browser
-`
-  }
-
-  throw Error(`unsupported language: ${language}`)
 }
 
 export default MonacoEditor
